fix mongodb nbRequests overflow

+++ modules/monitor/src/main/MongoStatus.scala
@@ -11,7 +11,7 @@ import lila.common.PimpedJson._
 private[monitor] case class MongoStatus(
   memory: Int = 0,
   connection: Int = 0,
-  query: Int = 0,
+  query: Long = 0,
   qps: Int = 0)
 
 private[monitor] object MongoStatus {
@@ -21,11 +21,11 @@ private[monitor] object MongoStatus {
   def apply(db: DB)(prev: MongoStatus): Fu[MongoStatus] =
     db.command(Status) map { bsonMap =>
       val status = JsObjectReader read BSONDocument(bsonMap)
-      val query = ~(status obj "network" flatMap (_ int "numRequests"))
+      val query = ~(status obj "network" flatMap (_ long "numRequests"))
       new MongoStatus(
         memory = ~(status \ "mem" \ "resident").asOpt[Int],
         connection = ~(status \ "connections" \ "current").asOpt[Int],
         query = query,
-        qps = query - prev.query)
+        qps = (query - prev.query).toInt)
     }
 }
